utils/pagination.py
from math import ceil

from rest_framework import pagination
from rest_framework.response import Response


class CustomPagination(pagination.PageNumberPagination):
    page_size_query_param = 'page_size'
    max_page_size = 16

    def get_paginated_response(self, data):
        return Response({
            'next': self.page.next_page_number() if self.page.has_next() else None,
            'previous': self.page.previous_page_number() if self.page.has_previous() else None,
            'count': self.page.paginator.count,
            'page_number': self.page.number,
            'all_pages': ceil(self.page.paginator.count / self.get_page_size(self.request)),
            'results': data
        })
