requirements.txt
Django==4.2.16
django-extensions==3.2.3
mysqlclient==2.2.4
djangorestframework==3.15.2
django-filter==24.3
